ROlled back changes involving bootstrap and other js.
Aloha has a clearly defined canonical source and simple ability to get it.  Bootstrap et al need
us to pull down a specific version, compile, diff compare and *then* use that - a longer process.
Since the files we want to use are in git anyway, leave them there till we fix the above