Add in own HTTPException handling
rather than rely on flask.abort, we raise an Exception class.  THis both
decouples us from Flask, and means we can raise an error in security,
and then let it bubble up, or catch and handle as needed.  For exmaple it
is valid for a folder you own to hold modules you dont, so an attempt to
read folder contents will result in a exception, but is not a 400 error.